moved altitude_rate.py and specific_energy_rate.py to non-flops-specific folder, upd flightpath.ode with ALTITUDE_RATE_MAX
