+++ aviary/mission/flops_based/ode/test/test_altitude_rate.py
@@ -2,7 +2,7 @@ import unittest
 
 import openmdao.api as om
 
-from aviary.mission.flops_based.ode.altitude_rate import AltitudeRate
+from aviary.mission.ode.altitude_rate import AltitudeRate
 from aviary.utils.test_utils.variable_test import assert_match_varnames
 from aviary.validation_cases.validation_data.flops_data.full_mission_test_data import \
     data
