+++ aviary/mission/gasp_based/ode/ascent_ode.py
@@ -13,8 +13,8 @@ from aviary.subsystems.aerodynamics.gasp_based.gaspaero import LowSpeedAero
 from aviary.subsystems.propulsion.propulsion_mission import PropulsionMission
 from aviary.variable_info.enums import AlphaModes
 from aviary.variable_info.variables import Aircraft, Dynamic, Mission
-from aviary.mission.flops_based.ode.specific_energy_rate import SpecificEnergyRate
-from aviary.mission.flops_based.ode.altitude_rate import AltitudeRate
+from aviary.mission.ode.specific_energy_rate import SpecificEnergyRate
+from aviary.mission.ode.altitude_rate import AltitudeRate
 
 
 class AscentODE(BaseODE):
