+++ aviary/mission/gasp_based/ode/descent_ode.py
@@ -19,8 +19,8 @@ from aviary.variable_info.enums import AnalysisScheme, SpeedType
 from aviary.variable_info.variables import Dynamic
 from aviary.subsystems.aerodynamics.aerodynamics_builder import AerodynamicsBuilderBase
 from aviary.subsystems.propulsion.propulsion_builder import PropulsionBuilderBase
-from aviary.mission.flops_based.ode.specific_energy_rate import SpecificEnergyRate
-from aviary.mission.flops_based.ode.altitude_rate import AltitudeRate
+from aviary.mission.ode.specific_energy_rate import SpecificEnergyRate
+from aviary.mission.ode.altitude_rate import AltitudeRate
 
 
 class DescentODE(BaseODE):
