+++ aviary/mission/ode/specific_energy_rate.py
